+++ packages/core/src/index.ts
@@ -1,5 +1,4 @@
 export * from './roxie';
-export * from './ajax';
 export * from './observable';
 export * from './data';
 export * from './types';